% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Creates `ssid` folder structure}
\usage{
create_folders(param = NULL)
}
\arguments{
\item{param}{Object of type `ssid_par` that bundles all `ssid` parameters,
including core model folders, alpha-3 country code and year and administrative unit
level at which the model is solved and type of
model.}
}
\description{
`create_folders` creates the folder structure that is needed to store the processed data for
`ssid`.
}
\details{
`create_folders` creates two folders in the `model_path`, set by the user:
mappings and processed_data (including subfolders). It copies a number of cvs files
into the mappings folder, which contain several data tables that are needed to run
the model and, if needed can be adjusted by the user.
}
\examples{
\dontrun{
create_folders(param)
}

}
