% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_template.R
\name{create_model_template}
\alias{create_model_template}
\title{Creates `ssid` model template}
\usage{
create_model_template(template_path = NULL)
}
\arguments{
\item{template_path}{folder where template scripts will be copied.}
}
\description{
`create_model_template` creates a folder structure with template R scripts to implement the various
steps to run `ssid`.
}
\details{
`create_model template()` creates several folders in the `template_path`, set by the user. The folders
contain template R scripts to implement `ssid`.

To run the template scripts, they have to be copied into an RStudio project. We recommend to
create such a project first and then use `create_model_template()` to add the scripts to
the RStudio project. Note that `create_model_template()` only copies script files if they do not exist
yet. In this way, the user cannot accidentally replace already modified scripts.
}
\examples{
\dontrun{
create_model_template("c:/temp")
}

}
