% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mids.R
\docType{package}
\name{mids-package}
\alias{mids}
\alias{mids-package}
\title{mids: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/michielvandijk/mids}
  \item Report bugs at \url{https://github.com/michielvandijk/mids/issues}
}

}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
