% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mids_par.R
\name{mids_par}
\alias{mids_par}
\title{Sets `mids` parameters}
\usage{
mids_par(
  model_path = NULL,
  db_path = NULL,
  db_path2 = NULL,
  iso3c = NULL,
  multi_country_run = FALSE,
  multi_country_region = NULL,
  adm_level = NULL,
  micro_year = NULL,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{model_path}{character. Folder where the model files are stored. Note
that R uses forward slash or double backslash to separate folder names.}

\item{db_path}{character. Folder where the input database is stored.}

\item{db_path2}{character. Optional second folder where the input database is
stored, for instance when household survey needs to be stored on a secured
location.}

\item{iso3c}{character. Vector with one or more (see `multi_country_run`)
three letter ISO 3166-1 alpha-3 country codes, also referred to as iso3c. A
list of country codes can be found in
[Wikipedia](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3).}

\item{multi_country_run}{logical. Whether the model is run for multiple
countries in a region for which survey information and constraint data is
available in single files. If TRUE, MIDS is looped over multiple countries
in a region. In this case the `iso3c` parameter must be a vector of country
codes, and the `region` parameter must be specified to determine which
countries to loop over. If FALSE, the model is run for a single country, and
the `iso3c` parameter must be a single country code. The default is FALSE.}

\item{multi_country_region}{character. Region or continent in which the
country is located required to loop the model over regional datasets (see
`multi_country_run`). The default is FALSE.}

\item{adm_level}{numeric. Administrative unit level at which the model is run.
This needs to correspond with the information in the shapefile of the
administrative units, which is needed as input.}

\item{micro_year}{numeric. Year for which the micro data (household survey) is
available.}

\item{start_year}{numeric. Start year of the simulation period.}

\item{end_year}{numeric. End year of the simulation period.}
}
\value{
mids_par object
}
\description{
`mids_par` sets all required parameters to run `mids`, including
 core model folders, country code and year.
}
\details{
`mids_par` creates an object of class `mids_par`, which bundles all
 required `mids` parameters set by the user: model and database folder, three
 digit country code, country name and continent.#'
 [countrycode::countrycode()] is used to determine the full country name,
 three digit country code and continent on the basis of the alpha-3 country
 code. This information is required to extract country specific information
 from several datasets.
}
\examples{
\dontrun{
mids_par(model_path = "C:/temp/mids_eth", db_path = "C:/temp/mids_db",
multi_country_run = TRUE, iso3c = "ETH", multi_country_region = "Africa",
adm_level = 2, micro_year = 2018, proj_year = 2021, start_year = 2018, end_year = 2050)
}
}
