% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mids.R
\name{summarise_mids}
\alias{summarise_mids}
\title{Summarise MIDS results}
\usage{
summarise_mids(original_data, mids_results)
}
\arguments{
\item{original_data}{Original input data frame.}

\item{mids_results}{Output from run_mids().}
}
\value{
A tibble with before, after and removed counts.
}
\description{
Summarises the number of original observations, retained MIDS individuals,
and removed repeat observations.
}
